Use hard line breaks for some properties
The operations, collectionOperations, resources, and rename properties
of service and resource shapes are now always on multiple lines if
provided and not empty.

For example:

```
service Foo {
    operations: [
        PutFoo
    ]
}
```

Includes a restructuring of smithy-syntax since it was getting
unwieldy.
